Refactor engine and move generation into smaller files
